role permission setting

+++ app/Http/Controllers/PermissionController.php
@@ -5,7 +5,8 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use Spatie\Permission\Models\Permission;
-
+use Gate;
+use Illuminate\Http\Response;
 class PermissionController extends Controller
 {
     /**
@@ -15,6 +16,8 @@ class PermissionController extends Controller
      */
     public function index()
     {   
+        abort_if(Gate::denies('permission_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $permissions = Permission::all();
 
         return view('admin.permissions.index', [
@@ -29,6 +32,8 @@ class PermissionController extends Controller
      */
     public function create() 
     {   
+        abort_if(Gate::denies('permission_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('admin.permissions.manage');
     }
 
@@ -40,6 +45,8 @@ class PermissionController extends Controller
      */
     public function store(Request $request)
     {   
+        abort_if(Gate::denies('permission_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $request->validate([
             'name' => 'required|unique:users,name'
         ]);
@@ -58,6 +65,8 @@ class PermissionController extends Controller
      */
     public function edit(Permission $permission)
     {
+        abort_if(Gate::denies('permission_update'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('admin.permissions.manage', [
             'permission' => $permission
         ]);
@@ -72,6 +81,8 @@ class PermissionController extends Controller
      */
     public function update(Request $request, Permission $permission)
     {
+        abort_if(Gate::denies('permission_update'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $request->validate([
             'name' => 'required|unique:permissions,name,'.$permission->id
         ]);
@@ -90,6 +101,8 @@ class PermissionController extends Controller
      */
     public function destroy(Permission $permission)
     {
+        abort_if(Gate::denies('permission_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $permission->delete();
 
         return redirect()->route('admin.permissions.index')
